backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch>=2.6.0
torchvision>=0.21.0
ultralytics>=8.1.0
python-multipart==0.0.6
pillow>=11.0.0
opencv-python==4.9.0.80
numpy<2.0.0
